🆙 update : button-slider  rename slider-control
