+++ libs/src/ui/module/slider-control/index.ts
@@ -1,3 +1,3 @@
 // @index(['./*.tsx', '!./*.stories.tsx'], f => `export * from '${f.path}'`)
-export * from './button-slider';
+export * from './slider-control';
 // @endindex
