+++ libs/src/ui/module/slider-control/slider-control.stories.tsx
@@ -4,7 +4,7 @@ import { ButtonSlider } from '@src/ui';
 import { PlusIcon, MinusIcon } from '@src/assets';
 
 export default {
-  title: 'Design System/Button-Slider',
+  title: 'Design System/Slider-Control',
   component: ButtonSlider,
   tags: ['autodocs'],
   argTypes: {
