+++ libs/src/ui/module/slider-control/slider-control.tsx
